# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        example: Invalid input
        type: string
    type: object
...
  handlers.ProjectResponse:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      description:
        example: Project description
        type: string
      id:
        example: 1
        type: integer
      name:
        example: My Project
        type: string
      owner:
        $ref: '#/definitions/handlers.UserResponse'
      owner_id:
        example: 1
        type: integer
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  handlers.RealTimeStatsResponse:
    properties:
      active_projects:
        example: 10
        type: integer
      connections_by_project:
        additionalProperties:
          type: integer
        type: object
      total_connections:
        example: 50
        type: integer
    type: object
...
  handlers.SuccessResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
  handlers.UserPresenceResponse:
    properties:
      is_online:
        example: true
        type: boolean
      last_seen:
        example: "2023-01-01T00:00:00Z"
        type: string
      project_id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
  handlers.UserResponse:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      email_verified:
        example: true
        type: boolean
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      last_login_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      name:
        example: John Doe
        type: string
      two_factor_enabled:
        example: false
        type: boolean
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  services.CreateProjectInput:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive task management API with real-time collaboration features
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Real Task API
  version: "1.0"
paths:
  /api/admin/realtime-stats:
    get:
      consumes:
      - application/json
      description: Get real-time system statistics including connection counts
      produces:
      - application/json
      responses:
        "200":
          description: Real-time statistics retrieved successfully
          schema:
            $ref: '#/definitions/handlers.RealTimeStatsResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get real-time statistics
      tags:
      - Real-time
...
  /api/projects:
    get:
      consumes:
      - application/json
      description: Get all projects for the authenticated user (owned + collaborated)
      produces:
      - application/json
      responses:
        "200":
          description: Projects retrieved successfully
          schema:
            items:
              $ref: '#/definitions/handlers.ProjectResponse'
            type: array
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Create a new project for the authenticated user
      parameters:
      - description: Project details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateProjectInput'
      produces:
      - application/json
      responses:
        "201":
          description: Project created successfully
          schema:
            $ref: '#/definitions/handlers.ProjectResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new project
      tags:
      - Projects
  /api/projects/{project_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing project (owner only)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Project deleted successfully
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Invalid project ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a project
      tags:
      - Projects
    get:
      consumes:
      - application/json
      description: Get details of a specific project by ID
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Project retrieved successfully
          schema:
            $ref: '#/definitions/handlers.ProjectResponse'
        "400":
          description: Invalid project ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a specific project
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Update an existing project's details
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Updated project details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateProjectInput'
      produces:
      - application/json
      responses:
        "200":
          description: Project updated successfully
          schema:
            $ref: '#/definitions/handlers.ProjectResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a project
      tags:
      - Projects
  /api/projects/{project_id}/connected-users:
    get:
      consumes:
      - application/json
      description: Get list of users currently connected to a specific project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Connected users retrieved successfully
          schema:
            items:
              $ref: '#/definitions/handlers.UserPresenceResponse'
            type: array
        "400":
          description: Invalid project ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get connected users for project
      tags:
      - Real-time
...
  /api/users/{user_id}/presence:
    get:
      consumes:
      - application/json
      description: Get presence information for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User presence retrieved successfully
          schema:
            $ref: '#/definitions/handlers.UserPresenceResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user presence
      tags:
      - Real-time
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"